Corrected issue with warped path placement at incorrect index
The issue stemmed from the encoding of certain characters requiring more
than 8 bits. As a result, the splitting process did not account for this
fact, leading to the insertion of the new path at the wrong location.
Notably, characters such as 'ü' caused the split index to shift,
resulting in the misplaced path.